move some nav stuff
